refactor: utils/配下のワークスペースを廃止しsrc/に統合
3つの独立パッケージ (notion-fetcher, notion-converter, notion-processor) を
src/配下のサブディレクトリに統合し、単一パッケージ構成に移行。

- utils/notion-fetcher/src/* -> src/fetcher/
- utils/notion-converter/src/* -> src/converter/
- utils/notion-processor/src/* -> src/processor/
- テストファイルをソースの隣にco-locate
- package.jsonからworkspacesを削除し依存関係を一元化
- root tsconfig.json, .eslintrc.jsonを統一作成
- 各パッケージの個別設定ファイル(package.json, tsconfig, eslint等)を削除
- utils/SPEC.mdをdocs/に移動
