+++ src/converter/image-handler.ts
