--- utils/notion-converter/.eslintrc.json
@@ -1,16 +1,0 @@
-{
-  "extends": [
-    "airbnb-base",
-    "airbnb-typescript/base",
-    "prettier"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": "./tsconfig.json"
-  },
-  "rules": {
-    "import/prefer-default-export": "off",
-    "no-console": ["warn", { "allow": ["warn", "error"] }],
-    "@typescript-eslint/no-explicit-any": "warn"
-  }
-}
