--- utils/notion-converter/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@document-hub/notion-converter",
-  "version": "1.0.0",
-  "description": "Notion記事をMarkdown形式に変換し、画像を適切に処理するコンバーターライブラリ",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "dev": "bun run src/index.ts",
-    "build": "bun build src/index.ts --outdir dist",
-    "test": "bun test",
-    "lint": "eslint src --ext .ts",
-    "lint:fix": "eslint src --ext .ts --fix",
-    "format": "prettier --write \"src/**/*.ts\"",
-    "type-check": "tsc --noEmit"
-  },
-  "dependencies": {
-    "notion-to-md": "^3.1.1",
-    "@notionhq/client": "^2.2.14"
-  },
-  "devDependencies": {
-    "@types/bun": "latest",
-    "typescript": "^5.3.3",
-    "eslint": "^8.56.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^17.1.0",
-    "@typescript-eslint/eslint-plugin": "^6.19.0",
-    "@typescript-eslint/parser": "^6.19.0",
-    "eslint-config-prettier": "^9.1.0",
-    "prettier": "^3.2.4",
-    "husky": "^8.0.3",
-    "lint-staged": "^15.2.0"
-  },
-  "lint-staged": {
-    "*.ts": [
-      "eslint --fix",
-      "prettier --write"
-    ]
-  }
-}
