--- utils/notion-fetcher/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "ESNext",
-    "lib": ["ES2022"],
-    "moduleResolution": "bundler",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "outDir": "./dist",
-    "rootDir": ".",
-    "types": ["bun-types"]
-  },
-  "include": ["src", "tests"],
-  "exclude": ["node_modules", "dist"]
-}
