--- utils/notion-processor/.env.example
@@ -1,2 +1,0 @@
-NOTION_API_KEY=your_notion_api_key
-NOTION_TEST_PAGE_ID=your_test_page_id
