--- utils/notion-processor/SPEC.md
@@ -1,1 +1,0 @@
-- 実際のNotion取得機能を実施するためのテスト追加
