--- utils/notion-processor/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@document-hub/notion-processor",
-  "version": "1.0.0",
-  "description": "Notion記事の取得とMarkdown変換を統合的に処理するライブラリ",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "type": "module",
-  "scripts": {
-    "dev": "bun run src/index.ts",
-    "build": "bun build src/index.ts --outdir dist --target node",
-    "test": "bun test tests/helpers.test.ts tests/integration.test.ts tests/processor.test.ts tests/unit.test.ts",
-    "test:all": "bun test",
-    "test:mocked": "bun test tests/helpers-mocked.test.ts",
-    "lint": "eslint src --ext .ts",
-    "lint:fix": "eslint src --ext .ts --fix",
-    "format": "prettier --write \"src/**/*.ts\"",
-    "type-check": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@document-hub/notion-fetcher": "workspace:*",
-    "@document-hub/notion-converter": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/bun": "latest",
-    "typescript": "^5.0.0",
-    "eslint": "^8.0.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint-plugin-import": "^2.29.0",
-    "prettier": "^3.0.0"
-  },
-  "peerDependencies": {
-    "@notionhq/client": "*"
-  },
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "keywords": [
-    "notion",
-    "markdown",
-    "converter",
-    "processor"
-  ],
-  "author": "",
-  "license": "MIT"
-}
